+++ main-application/src/main/java/event/gps/GPSConnect.java
@@ -1,4 +1,4 @@
-package event.GPS;
+package event.gps;
 
 public class GPSConnect {
         private String phase;
