+++ main-application/src/main/java/event/gps/GPSOn.java
@@ -1,4 +1,4 @@
-package event.GPS;
+package event.gps;
 
 public class GPSOn {
         private String phase;
