+++ main-application/src/main/java/event/gps/GPSReceive.java
@@ -1,4 +1,4 @@
-package event.GPS;
+package event.gps;
 
 public class GPSReceive {
     private String phase;
