+++ main-application/src/main/java/event/satcom/SatComConnect.java
@@ -1,4 +1,4 @@
-package event.SatCom;
+package event.satcom;
 
 public class SatComConnect {
     private String phase;
