+++ main-application/src/test/java/SatComTest.java
@@ -55,10 +55,8 @@ public class SatComTest {
         try {
             Method method1 = componentPort.getClass().getDeclaredMethod("on");
             method1.invoke(componentPort);
-            Method offMethod = componentPort.getClass().getDeclaredMethod("connect", String.class);
-            String satellite = "Ares4";
-            String frequency = "450Hz";
-            boolean connectResult = (boolean)offMethod.invoke(componentPort, new Object[]{satellite}, new Object[]{frequency});
+            Method offMethod = componentPort.getClass().getDeclaredMethod("connect", String.class, String.class);
+            boolean connectResult = (boolean)offMethod.invoke(componentPort, new Object[]{"Ares4", "450Hz"});
             assertEquals(true, connectResult);
         } catch (Exception e) {
             System.out.println(e.getMessage());
@@ -71,8 +69,8 @@ public class SatComTest {
         try {
             Method method1 = componentPort.getClass().getDeclaredMethod("on");
             method1.invoke(componentPort);
-            Method method2 = componentPort.getClass().getDeclaredMethod("connect", String.class);
-            method2.invoke(componentPort,new Object[]{"Ares3"} );
+            Method method2 = componentPort.getClass().getDeclaredMethod("connect", String.class, String.class);
+            method2.invoke(componentPort,new Object[]{"Ares3", "350Hz"} );
             Method offMethod = componentPort.getClass().getDeclaredMethod("receive");
             String receiveResult = (String)offMethod.invoke(componentPort);
             assertEquals("No weather warning ahead", receiveResult);
